add notification trigger

+++ src/controllers/post.controller.ts
@@ -1,5 +1,7 @@
 import { Controller, Post, Get, Patch, Delete, Body, Param, Query, ParseIntPipe, UseGuards, 
-    UseInterceptors, ParseFilePipeBuilder, UploadedFiles, StreamableFile, Res } from '@nestjs/common';
+    UseInterceptors, ParseFilePipeBuilder, UploadedFiles, StreamableFile, Res, 
+    Request,
+    ForbiddenException} from '@nestjs/common';
 import { FilesInterceptor } from '@nestjs/platform-express';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { Role, Roles } from 'src/auth/role.decorator';
@@ -21,10 +23,12 @@ export class PostController {
     @Post()
     @UseInterceptors(FilesInterceptor('files'))
     async addPost(@UploadedFiles(new ParseFilePipeBuilder().addMaxSizeValidator(null).build({fileIsRequired: false})) files: Array<Express.Multer.File>,
-                  @Body(new ParseFormDataPipe, new ValidationPipe) postDto: PostDto){
+                  @Body(new ParseFormDataPipe, new ValidationPipe) postDto: PostDto, @Request() req){
+        const user = JSON.parse(req.user.profile)
+        postDto.userId = user.id;
         const newPost = await this.postService.Add(postDto);
         if(files)
-            await this.imgService.AddPostImage(newPost.id, files)
+            await this.imgService.AddPostImage(newPost.id, user.id, files)
         return {message: 'Post uploaded'};
     }
 
@@ -54,21 +58,28 @@ export class PostController {
     @Patch(':id')
     @UseInterceptors(FilesInterceptor('files'))
     async updatePost(@UploadedFiles(new ParseFilePipeBuilder().build({fileIsRequired: false})) files: Array<Express.Multer.File>, 
-        @Body(new ParseFormDataPipe, new ValidationPipe()) post: PostDto, @Param('id', ParseIntPipe) postId: number){
-
-        const message = await this.postService.UpdatePost(postId, post);
+        @Body(new ParseFormDataPipe, new ValidationPipe()) postDto: PostDto, @Param('id', ParseIntPipe) postId: number, @Request() req){
+        const user = JSON.parse(req.user.profile);
+        const post = await this.postService.FindOneById(postId);
+        if(post.user.id !== user.id)
+            throw new ForbiddenException("Cannot edit other's post");
+        const message = await this.postService.UpdatePost(postId, postDto);
         if(files)
             await this.imgService.DeletePostImages(postId);
-            await this.imgService.AddPostImage(postId, files);
+            await this.imgService.AddPostImage(postId, user.id, files);
         return message;
     }
 
 
     @Delete(':id')
-    async deletePost(@Param('id', ParseIntPipe) postId: number, @Query('userId', ParseIntPipe)userId: number){
+    async deletePost(@Param('id', ParseIntPipe) postId: number, @Request() req){
+        const user = JSON.parse(req.user.profile);
+        const post = await this.postService.FindOneById(postId);
+        if(post.user.id !== user.id)
+            throw new ForbiddenException("Cannot delete other's post !");
         await this.imgService.DeletePostImages(postId);
         await this.activityService.DeletePostComments(postId);
         await this.activityService.DeletePostActivities(postId);
-        return await this.postService.DeletePost(postId, userId);
+        return await this.postService.DeletePost(postId);
     }
 }
